+++ lib-opal/opal/rspec/browser.rb
@@ -1,5 +1,9 @@
 require 'opal/rspec/formatter/browser_formatter'
 
 RSpec.configure do |config|
-  config.default_formatter = ::Opal::RSpec::BrowserFormatter
+  if OPAL_PLATFORM.nil?
+    # We want the browser formatter ONLY for the real browser, not
+    # our headless browser runners.
+    config.default_formatter = ::Opal::RSpec::BrowserFormatter
+  end
 end